remote_travel_agents/train_agent/pyproject.toml
[project]
name = "train-agent"
version = "0.1.0"
description = "A2A Train Reservation Agent using LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langgraph>=0.3.34",
    "pydantic>=2.10.6",
    "langchain-google-vertexai>=2.0.21",
    "dotenv>=0.9.9",
    "jwcrypto>=1.5.6",
    "pyjwt>=2.10.1",
    "sse-starlette>=2.3.3",
    "uvicorn>=0.34.2",
    "a2a-sdk>=0.2.16",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
